studio 19 done

+++ Studio19/UnitTest/UnitTest.cpp
@@ -50,7 +50,7 @@ namespace UnitTest
 			std::vector<char> v = { 'h', 'i' };
 			Assert::AreEqual(t.write(v), 0);
 			std::vector<char> contents = t.read();
-			Assert::AreEqual(t.getSize(), v.size());
+			Assert::AreEqual(t.getSize(), static_cast<unsigned int>( v.size()));
 			Assert::AreEqual(contents.size(), v.size());
 			for (size_t i = 0; i < contents.size(); ++i) {
 				Assert::AreEqual(v[i], contents[i]);
@@ -253,6 +253,8 @@ public:
 		Assert::IsTrue(isNull);
 	}
 	};
+
+	
 	TEST_CLASS(basicDisplayVisitor) {
 public:
 	TEST_METHOD(visitTextFile) { // tests output of basic display visitor for a text file, expects the contents of the file
@@ -366,4 +368,5 @@ public:
 		cout.rdbuf(backup);
 	}
 	};
+	
 }
